Removed video mode detection + regressions
As it turns out, detecting video mode within 446 B is not a trivial
task. Furthermore, video modes with more than 80 columns are non-
standard, and are not guaranteed to exist on all platforms.

With that said, future iterations might attempt to toggle the font
being used in text mode - while still using 80-column mode.

The sources compile, but have never been tested.

Signed-off-by: John Isaac Calderon <ogclapdrop@gmail.com>