Used aliases and build dir in Makefile + fixes
The Makefile should be somewhat safer now, but not quite.

Since we're now using a variable to specify the build
directory, there is a risk that the symbol name in
'boot64_wrap.bin' is inconsistent - leading to intractable
bugs that only vanish when default settings are used. The
changes in the linker script reflect that risk.

Signed-off-by: John Isaac Calderon <ogclapdrop@gmail.com>