Added HAL basics + evaluating PIC options
A macro to quickly define functions and statics within
custom sections has been implemented. This is later used
to define the HAL and its associated vector table.

It is being evaluated whether the 64-bit part of 'boot1'
can be made position-independent, so that the HAL can be
reused by dependents like a kernel.

Some project restructuring was performed.

Signed-off-by: John Isaac Calderon <ogclapdrop@gmail.com>