Added FS image generation + reorganization
The build process has been revised so that it generates a bootable
disk image containing a FAT16 filesystem - which itself contains
the second-stage bootloader `boot1.bin`.

The VBR is still a work in progress - and will be the main focus
of the `experimental` branch.

The project has been reorganized, so that scripts are located in a
separate `scripts/` directory. The `scripts/patch_vbr.sh` helper
script patches the built VBR into an existing volume without
modifying the existing EBPB.

Signed-off-by: John Isaac Calderon <ogclapdrop@gmail.com>