Finalized VBR + changed stubs for proper loading
The VBR wasn't entirely finished in the previous iterations, as
parts of the logic was missing, which led to erroneous loads
and unplanned panics. It's patchwork - like everything else, but
it works for the time being.

The project structure has also been altered, so that shared
constants are defined elsewhere.

The stubs, as well as related files, have been altered to
accomodate this radical change in the boot sequence.

Signed-off-by: John Isaac Calderon <ogclapdrop@gmail.com>