Finalized VBR + changed stubs for proper loading
The VBR wasn't entirely finished in the previous iterations, as
parts of the logic was missing, which led to erroneous loads and
unplanned panics. Like everything else, it is patchwork, but it
works for the time being.

1. The reading routine in the VBR should now account for
segmentation. This is one of the reasons why the routine
only reads one sector per loop cycle.

2. The stubs have been refactored and restructured, with `stub16.asm`
containing 16-bit code and `stub32.asm` containing 32-bit code. They
will be combined together by the assembler, but having them separated
should facilitate situational awareness.

3. Shared constants are now defined elsewhere: affected source files
have been modified to accomodate this radical change.

4. The CPUID check has been moved from protected mode to real mode,
and is more direct than checking MSRs for capabilities. Further
checks are still performed in 32-bit PM to maintain readability and
save space on instructions. (32-bit operations require the use of
prefixes `0x66` and `0x67`)

Signed-off-by: John Isaac Calderon <ogclapdrop@gmail.com>