Correctness, doc supplements & build process
(Correctness & doc supplements)
1. There were some missing documentation comments - these
should be added now.

2. Some instructions and procedures had to be added or modified
in order for them to be "correct" on the 8086/80186, among them
being the high nibble check on FLAGS - as bits 12-14 should be
sticky on processors newer than the 80186. This check will be
used to prevent booting on obviously outdated platforms.

(Changed build process to use `elf_x86_64`)
1. This turns out to be a much cleaner approach than wrapping
pre- compiled 64-bit routines in a linkable format, particularly
if the routines use absolute addresses internally.

2. Due to how the project was restructured earlier, this has
finally become possible. It builds and runs like usual, and
hard-coded accesses are resolved as expected.

Signed-off-by: John Isaac Calderon <ogclapdrop@gmail.com>