Made HAL VT relocatable + removed redundancies
The vector table struct should now be able to translate function
pointers internally, so that they point to methods with respect
to the VT's new base - instead of the old base, or +0x0. This
approach demands discipline: the VT should ideally be in the same
section as the methods it is pointing to.

The `dispatch` method no longer returns lock handles. Related
structs have therefore been removed. Atomic access is instead
handled internally. This, of course, demands discipline: all
clients wishing to access the VT must use the same API and
ABI - but it can be done...

Signed-off-by: John Isaac Calderon <ogclapdrop@gmail.com>