Established shared resources module
As per decision 2025-08-30, part 2, platform-specific
definitions must be separated from shared defintions,
and must be guarded off whenever possible.

For the sake of completeness, the example code in
`kern/lib.rs` has been replaced with a proper
template.

A proposed layout for the module, which is being implemented
by this commit, has been immortalized in decision 2025-09-01,
part 1.

(revision 2025-11-29) Decision 2025-11-29 is in force.

Signed-off-by: John Isaac Calderon <ogclapdrop@gmail.com>