Revised I/O traits implementation
The order of prerequisites has been inversed, so that `Read` and
`Write` derive from `CharDevice`, rather than the other way round.

This is because `Read` and `Write` are generic traits, whereas
`CharDevice` is a special trait (marks a type as a character
device).

Also moved I/O traits, which were previously under `shared::traits`,
to `shared::io`.

Signed-off-by: John Isaac Calderon <ogclapdrop@gmail.com>