Revised build process for complete codegen
Since the bootloader relies on implicit panicking, definitions
from `core` must be present at link time. This turned out to not
be the case.

The solution is swift: compile the Rust component into a static
archive (which should all definitions), and then remove unused
references at link time.

Offsets stored in the headers are calculated by the assembler,
rather than by the linker.

Signed-off-by: John Isaac Calderon <ogclapdrop@gmail.com>